Revert "First-pass at racial data collection."
This reverts commit 961895e2c1f9d609a2f28eeaef6984b2d84a73cf.
