+++ public/controllers/mainController.js
@@ -18,7 +18,7 @@ app.controller('mainController', ['$scope', '$http', function($scope, $http) {
             alignment: "",
             deity: "",
 
-            ddrace: "Dwarf",
+            ddrace: "",
             abilityBonus1: "",
             abilityBonus2: "",
             skillBonus1: "",
